requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
SQLAlchemy==2.0.4
twilio==7.16.3
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.14.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
SQLAlchemy==2.0.4
twilio==7.16.3
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.14.0